+++ src/components/library/RiggedHand/RiggedHand.cpp
@@ -223,7 +223,7 @@ void RiggedHand::DrawContents(RenderState& renderState) const {
       auto locations = std::make_tuple(positionAddr, normalAddr, texcoordAddr, boneWeightsAddr, boneIndicesAddr);
       section->getVboMesh().Enable(locations);
 
-      const int numIndices = section->getIndices().Size();
+      const int numIndices = static_cast<int>(section->getIndices().Size());
       section->getIndices().Bind();
       GL_THROW_UPON_ERROR(glDrawElements(GL_TRIANGLES, numIndices, GL_UNSIGNED_INT, 0));
       section->getIndices().Unbind();
